Create image::view module, move view-related types into it
