+++ vulkano/src/image/attachment.rs
@@ -23,18 +23,18 @@ use format::FormatDesc;
 use format::FormatTy;
 use image::sys::ImageCreationError;
 use image::sys::UnsafeImage;
-use image::sys::UnsafeImageView;
 use image::traits::ImageAccess;
 use image::traits::ImageClearValue;
 use image::traits::ImageContent;
-use image::traits::ImageViewAccess;
+use image::view::ImageViewAccess;
+use image::view::ImageViewDimensions;
+use image::view::ImageViewType;
+use image::view::UnsafeImageView;
 use image::ImageCreateFlags;
 use image::ImageDimensions;
 use image::ImageInner;
 use image::ImageLayout;
 use image::ImageUsage;
-use image::ImageViewDimensions;
-use image::ImageViewType;
 use memory::pool::AllocFromRequirementsFilter;
 use memory::pool::AllocLayout;
 use memory::pool::MappingRequirement;
