+++ vulkano/src/image/immutable.rs
@@ -29,15 +29,15 @@ use format::Format;
 use format::FormatDesc;
 use image::sys::ImageCreationError;
 use image::sys::UnsafeImage;
-use image::sys::UnsafeImageView;
 use image::traits::ImageAccess;
 use image::traits::ImageContent;
-use image::traits::ImageViewAccess;
+use image::view::ImageViewAccess;
+use image::view::ImageViewDimensions;
+use image::view::UnsafeImageView;
 use image::ImageCreateFlags;
 use image::ImageInner;
 use image::ImageLayout;
 use image::ImageUsage;
-use image::ImageViewDimensions;
 use image::MipmapsCount;
 use instance::QueueFamily;
 use memory::pool::AllocFromRequirementsFilter;
