+++ vulkano/src/image/sys.rs
@@ -7,10 +7,10 @@
 // notice may not be copied, modified, or distributed except
 // according to those terms.
 
-//! Low-level implementation of images and images views.
+//! Low-level implementation of images.
 //!
-//! This module contains low-level wrappers around the Vulkan image and image view types. All
-//! other image or image view types of this library, and all custom image or image view types
+//! This module contains low-level wrappers around the Vulkan image types. All
+//! other image types of this library, and all custom image types
 //! that you create must wrap around the types in this module.
 
 use smallvec::SmallVec;
@@ -33,7 +33,6 @@ use image::ImageAspect;
 use image::ImageCreateFlags;
 use image::ImageDimensions;
 use image::ImageUsage;
-use image::ImageViewType;
 use image::MipmapsCount;
 use memory::DeviceMemory;
 use memory::DeviceMemoryAllocError;
@@ -1012,237 +1011,6 @@ pub struct LinearLayout {
     /// Number of bytes between two texels or two blocks in adjacent depth layers. This value is
     /// undefined for images that are not three-dimensional.
     pub depth_pitch: usize,
-}
-
-pub struct UnsafeImageView {
-    view: vk::ImageView,
-    device: Arc<Device>,
-    usage: vk::ImageUsageFlagBits,
-    identity_swizzle: bool,
-    format: Format,
-}
-
-impl UnsafeImageView {
-    /// See the docs of new().
-    pub unsafe fn raw(
-        image: &UnsafeImage,
-        ty: ImageViewType,
-        mipmap_levels: Range<u32>,
-        array_layers: Range<u32>,
-    ) -> Result<UnsafeImageView, OomError> {
-        let vk = image.device.pointers();
-
-        assert!(mipmap_levels.end > mipmap_levels.start);
-        assert!(mipmap_levels.end <= image.mipmaps);
-        assert!(array_layers.end > array_layers.start);
-        assert!(array_layers.end <= image.dimensions.array_layers());
-        assert!(
-            (
-                vk::IMAGE_USAGE_SAMPLED_BIT
-                    | vk::IMAGE_USAGE_STORAGE_BIT
-                    | vk::IMAGE_USAGE_COLOR_ATTACHMENT_BIT
-                    | vk::IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT
-                    | vk::IMAGE_USAGE_INPUT_ATTACHMENT_BIT
-                    | vk::IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT
-                // TODO | vk::IMAGE_USAGE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR
-                // TODO | vk::IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT
-            ) & image.usage.to_usage_bits()
-                != 0
-        );
-
-        let aspect_mask = match image.format.ty() {
-            FormatTy::Float | FormatTy::Uint | FormatTy::Sint | FormatTy::Compressed => {
-                vk::IMAGE_ASPECT_COLOR_BIT
-            }
-            FormatTy::Depth => vk::IMAGE_ASPECT_DEPTH_BIT,
-            FormatTy::Stencil => vk::IMAGE_ASPECT_STENCIL_BIT,
-            FormatTy::DepthStencil => vk::IMAGE_ASPECT_DEPTH_BIT | vk::IMAGE_ASPECT_STENCIL_BIT,
-            // Not yet supported --> would require changes to ImmutableImage API :-)
-            FormatTy::Ycbcr => panic!(),
-        };
-
-        let view_type = match (
-            image.dimensions(),
-            image.create_flags().cube_compatible,
-            ty,
-            array_layers.end - array_layers.start,
-        ) {
-            (ImageDimensions::Dim1d { .. }, _, ImageViewType::Dim1d, 1) => vk::IMAGE_VIEW_TYPE_1D,
-            (ImageDimensions::Dim1d { .. }, _, ImageViewType::Dim1dArray, _) => {
-                vk::IMAGE_VIEW_TYPE_1D_ARRAY
-            }
-            (ImageDimensions::Dim2d { .. }, _, ImageViewType::Dim2d, 1) => vk::IMAGE_VIEW_TYPE_2D,
-            (ImageDimensions::Dim2d { .. }, _, ImageViewType::Dim2dArray, _) => {
-                vk::IMAGE_VIEW_TYPE_2D_ARRAY
-            }
-            (ImageDimensions::Dim2d { .. }, true, ImageViewType::Cubemap, 6) => {
-                vk::IMAGE_VIEW_TYPE_CUBE
-            }
-            (ImageDimensions::Dim2d { .. }, true, ImageViewType::CubemapArray, n) => {
-                assert_eq!(n % 6, 0);
-                vk::IMAGE_VIEW_TYPE_CUBE_ARRAY
-            }
-            (ImageDimensions::Dim3d { .. }, _, ImageViewType::Dim3d, _) => vk::IMAGE_VIEW_TYPE_3D,
-            _ => panic!(),
-        };
-
-        let view = {
-            let infos = vk::ImageViewCreateInfo {
-                sType: vk::STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO,
-                pNext: ptr::null(),
-                flags: 0, // reserved
-                image: image.internal_object(),
-                viewType: view_type,
-                format: image.format as u32,
-                components: vk::ComponentMapping {
-                    r: 0,
-                    g: 0,
-                    b: 0,
-                    a: 0,
-                }, // FIXME:
-                subresourceRange: vk::ImageSubresourceRange {
-                    aspectMask: aspect_mask,
-                    baseMipLevel: mipmap_levels.start,
-                    levelCount: mipmap_levels.end - mipmap_levels.start,
-                    baseArrayLayer: array_layers.start,
-                    layerCount: array_layers.end - array_layers.start,
-                },
-            };
-
-            let mut output = MaybeUninit::uninit();
-            check_errors(vk.CreateImageView(
-                image.device.internal_object(),
-                &infos,
-                ptr::null(),
-                output.as_mut_ptr(),
-            ))?;
-            output.assume_init()
-        };
-
-        Ok(UnsafeImageView {
-            view,
-            device: image.device.clone(),
-            usage: image.usage.to_usage_bits(),
-            identity_swizzle: true, // FIXME:
-            format: image.format,
-        })
-    }
-
-    /// Creates a new view from an image.
-    ///
-    /// Note that you must create the view with identity swizzling if you want to use this view
-    /// as a framebuffer attachment.
-    ///
-    /// # Panic
-    ///
-    /// - Panics if `mipmap_levels` or `array_layers` is out of range of the image.
-    /// - Panics if the view types doesn't match the dimensions of the image (for example a 2D
-    ///   view from a 3D image).
-    /// - Panics if trying to create a cubemap with a number of array layers different from 6.
-    /// - Panics if trying to create a cubemap array with a number of array layers not a multiple
-    ///   of 6.
-    /// - Panics if the device or host ran out of memory.
-    /// - Panics if the image is a YcbCr image, since the Vulkano API is not yet flexible enough to
-    ///   specify the aspect of image.
-    #[inline]
-    pub unsafe fn new(
-        image: &UnsafeImage,
-        ty: ImageViewType,
-        mipmap_levels: Range<u32>,
-        array_layers: Range<u32>,
-    ) -> UnsafeImageView {
-        UnsafeImageView::raw(image, ty, mipmap_levels, array_layers).unwrap()
-    }
-
-    #[inline]
-    pub fn format(&self) -> Format {
-        self.format
-    }
-
-    #[inline]
-    pub fn usage_transfer_source(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_TRANSFER_SRC_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_transfer_destination(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_TRANSFER_DST_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_sampled(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_SAMPLED_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_storage(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_STORAGE_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_color_attachment(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_COLOR_ATTACHMENT_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_depth_stencil_attachment(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_transient_attachment(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT) != 0
-    }
-
-    #[inline]
-    pub fn usage_input_attachment(&self) -> bool {
-        (self.usage & vk::IMAGE_USAGE_INPUT_ATTACHMENT_BIT) != 0
-    }
-}
-
-unsafe impl VulkanObject for UnsafeImageView {
-    type Object = vk::ImageView;
-
-    const TYPE: vk::ObjectType = vk::OBJECT_TYPE_IMAGE_VIEW;
-
-    #[inline]
-    fn internal_object(&self) -> vk::ImageView {
-        self.view
-    }
-}
-
-impl fmt::Debug for UnsafeImageView {
-    #[inline]
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        write!(fmt, "<Vulkan image view {:?}>", self.view)
-    }
-}
-
-impl Drop for UnsafeImageView {
-    #[inline]
-    fn drop(&mut self) {
-        unsafe {
-            let vk = self.device.pointers();
-            vk.DestroyImageView(self.device.internal_object(), self.view, ptr::null());
-        }
-    }
-}
-
-impl PartialEq for UnsafeImageView {
-    #[inline]
-    fn eq(&self, other: &Self) -> bool {
-        self.view == other.view && self.device == other.device
-    }
-}
-
-impl Eq for UnsafeImageView {}
-
-impl Hash for UnsafeImageView {
-    #[inline]
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.view.hash(state);
-        self.device.hash(state);
-    }
 }
 
 #[cfg(test)]
